feat(flake.nix): init flake.{nix, lock}.
This includes a basic development shell for programming on linux.
NOTE: This can support macos, as well as other linux systems that aren't
x86_64-linux (like arm). It could also support building the project itself,
but this commit includes nothing beyond a basic way to setup a devenv
on linux.