refactor(scheduler): make SystemState be passed in as a parameter.
We do this because multiple schedulers might be running at the same
time, and the virtue of simply the scheduler running doesnn't mean
nothing else has in that time period. That is to say: calling tick() on
a cpu scheduler shouldn't advance system time if the I/O scheduler could
still do something, or one of the other cores.