refactor(scheduler): make scheduler return state about the `tick`
This is needed since in the future when we have both a CPU and an I/O
scheduler, a process needs to be able to swap between the two. That
means `tick` needs to give back the process when it's CPU Burst time is
done, so someone else can take over.